% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assemblage-package.R
\docType{package}
\name{Assemblage-package}
\alias{Assemblage}
\alias{Assemblage-package}
\title{Assemblage: Assemblage Regression}
\description{
The concept of assemblage originates from the research paper titled "Maximally Forward-Looking Core Inflation" (2024) authored by Philippe Goulet Coulombe, Karin Klieber, Christophe Barrette, and Maximilian Göbel. Assemblage Regression represents a specialized form of generalized nonnegative ridge regression, designed to optimize the weights of subcomponents to maximize the predictive capability of the aggregate.
}
\keyword{internal}
