% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage_v240228.R
\name{assemblage.estimation.RW}
\alias{assemblage.estimation.RW}
\title{assemblage.estimation.RW}
\usage{
assemblage.estimation.RW(
  y,
  x,
  bench,
  assemblage.info,
  moving.average,
  end.in.sample,
  window.size,
  step.estim.window,
  standardize.values,
  horizon.gap,
  coef.sum.1,
  lambda.grid.C = c(),
  ncores,
  model.select,
  all.sample.fit,
  Progression
)
}
\arguments{
\item{y}{Target vector}

\item{x}{The raw x matrix}

\item{bench}{The raw Benchmarks matrix}

\item{assemblage.info}{List containing the current assemblage information}

\item{moving.average}{This parameter governs the transformation of the x matrix. If set to 1, the x matrix will be retained for Month over Month analysis. Setting it to 3 corresponds to Quarter over Quarter analysis, while 12 signifies Year over Year analysis.}

\item{end.in.sample}{The observation when the in-sample estimation end}

\item{window.size}{Size of the rolling window}

\item{step.estim.window}{Estimating the rolling window every ... observations}

\item{standardize.values}{Standardize the response variables}

\item{horizon.gap}{The number of observations to be excluded to ensure that there is no overlapping information within the out-of-sample dataset}

\item{coef.sum.1}{Does the coefficients need to sum to one}

\item{lambda.grid.C}{Personalize lambda grid}

\item{ncores}{Number of cores to run the code, helps for glmnet function. (default is 1)}

\item{model.select}{Select the model to run}

\item{all.sample.fit}{Do an estimation using all the data set as training (default is FALSE)}

\item{Progression}{Give feedback on the estimation. (default 0)}
}
\description{
Function estimating the rolling window assemblage
}
