% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage_v240228.R
\name{nonneg.ridge.mean}
\alias{nonneg.ridge.mean}
\title{nonneg.ridge.mean}
\usage{
nonneg.ridge.mean(
  y.in,
  x.in,
  standardize.values,
  lambda.grid.C = c(),
  ncores = 1
)
}
\arguments{
\item{y.in}{The target variable}

\item{x.in}{The features matrix}

\item{standardize.values}{Standardize the response variables}

\item{lambda.grid.C}{Personalize lambda grid}

\item{ncores}{Number of cores to run the code, helps for glmnet function. (default is 1)}
}
\description{
Function used to find the coefficients of the blend in the nonnegative ridge regression
}
