% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage_v240228.R
\name{nonneg.ridge}
\alias{nonneg.ridge}
\title{nonneg.ridge}
\usage{
nonneg.ridge(y.in, x.in, standardize.values)
}
\arguments{
\item{y.in}{The target variable}

\item{x.in}{The features matrix}

\item{standardize.values}{Standardize the response variables}
}
\description{
Function used to find the coefficients in the nonnegative ridge regression
}
