% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage_v240228.R
\name{RMSE.function}
\alias{RMSE.function}
\title{RMSE.function}
\usage{
RMSE.function(Prediction, Start.POOS = c(), bench = c())
}
\arguments{
\item{Prediction}{The prediction matrix}

\item{Start.POOS}{Where the POOS start}

\item{bench}{The benchmark matrix}
}
\description{
Function to calculate the RMSE
}
