% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage_v240228.R
\name{Time.function}
\alias{Time.function}
\title{Time.function}
\usage{
Time.function(ooo, end.ooo, start.time)
}
\arguments{
\item{ooo}{Iteration of the expanding window}

\item{end.ooo}{Number of iterations}

\item{start.time}{When the expanding window started}
}
\description{
Function to have a time estimation for the expanding window
}
