% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage_v240228.R
\name{weight.transformation}
\alias{weight.transformation}
\title{weight.transformation}
\usage{
weight.transformation(x.weight, x.comps, moving.average)
}
\arguments{
\item{x.weight}{The raw weight matrix can be left empty, or it can be a vector or a matrix. If used, its components should align in the same order as those in 'x'.}

\item{x.comps}{The components matrix}

\item{moving.average}{This parameter governs the transformation of the x matrix. If set to 1, the x matrix will be retained for Month over Month analysis. Setting it to 3 corresponds to Quarter over Quarter analysis, while 12 signifies Year over Year analysis.}
}
\description{
Function used to find and calculate the components weights in the right moving average unit
}
