% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage_v240228.R
\name{x.transformation}
\alias{x.transformation}
\title{x.transformation}
\usage{
x.transformation(x, moving.average)
}
\arguments{
\item{x}{The matrix we need to transform and rank (Should be in month-over-month format)}

\item{moving.average}{This parameter governs the transformation of the x matrix. If set to 1, the x matrix will be retained for Month over Month analysis. Setting it to 3 corresponds to Quarter over Quarter analysis, while 12 signifies Year over Year analysis.}
}
\description{
Function used to rank and transform the x matrix into the right moving average unit
}
